requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Google Search API
google-api-python-client==2.108.0

# Document Processing
pdfminer.six==20221105
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
pytesseract==0.3.10

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Email
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Browser Automation (Optional - for future spoofing module)
selenium==4.15.2
webdriver-manager==4.0.1
